feat: 分离组件化，添加合作伙伴模块
